refactor: config module
